🎁 Tweaking the Pokemon in Timeline
* Adjusted max width
* Refactored Box components
* Hide BadgeBox if no badges
