+++ styles/Pokemon.module.scss
@@ -35,8 +35,7 @@
     position: relative;
 
     img {
-      max-width: 100%;
-      max-height: 100%;
+      position: absolute;
     }
   }
 }
