+++ styles/Summary.module.scss
@@ -129,7 +129,6 @@
     align-items: flex-end;
     .gravePokemon {
       flex-direction: row-reverse;
-      justify-content: flex-end;
     }
   }
